Adapt the Environment module to the new config/runtime.exs setup
